Removed the new file, left the other one with the franc library
